EF Core Code first

+++ EFDbFirstDemo/EFDbFirstDemo/Program.cs
@@ -16,28 +16,32 @@ namespace EFDbFirstDemo
             Console.WriteLine(blog?.Title);
 
             //add new entry
-            //Blog blogToAdd = new Blog();
-            //blogToAdd.Title = "Satya Blog";
-            //blogToAdd.Author = "Satya";
-            //blogToAdd.CreatedDt = DateTime.Now;
-            //blogToAdd.CreatedBy = "Satya";
+            Blog blogToAdd = new Blog();
+            blogToAdd.Title = "Satya Blog";
+            blogToAdd.Author = "Satya";
+            blogToAdd.CreatedDt = DateTime.Now;
+            blogToAdd.CreatedBy = "Satya";
 
-            //_context.Blogs.Add(blogToAdd);
-            //_context.SaveChanges();
+            _context.Blogs.Add(blogToAdd);
+            _context.SaveChanges();
 
             //Update script
-            //Blog blogToUpdate = _context.Blogs.Where(b => b.Id == 3).FirstOrDefault();
-            //blogToUpdate.Title = "blog updated - .net core";
-            //blogToUpdate.UpdatedDt = DateTime.Now;
-            //blogToUpdate.UpdatedBy = "System";
-            //_context.Update(blogToUpdate);
-            //_context.SaveChanges();
+            Blog blogToUpdate = _context.Blogs.Where(b => b.Id == 2).FirstOrDefault();
+            blogToUpdate.Title = "blog updated - .net core";
+            blogToUpdate.UpdatedDt = DateTime.Now;
+            blogToUpdate.UpdatedBy = "System";
+            _context.Update(blogToUpdate);
+            _context.SaveChanges();
 
 
             //Delete script
             Blog blogToRemove = _context.Blogs.Where(b => b.Id == 3).FirstOrDefault();
-            _context.Blogs.Remove(blogToRemove);
-            _context.SaveChanges();
+            if(blogToRemove != null)
+            {
+                _context.Blogs.Remove(blogToRemove);
+                _context.SaveChanges();
+            }
+           
 
 
         }
